Move books.js and categories.js to components folder
