+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import React from 'react';
 import { HashRouter as Router, Route, Switch } from 'react-router-dom';
-import Books from './redux/books/books';
-import Categories from './redux/categories/categories';
+import Books from './components/books';
+import Categories from './components/categories';
 import Navbar from './components/Navbar';
 
 function App() {
